thorw key error if use_cls_token is not set as default value.
