+++ rasa/nlu/tokenizers/mitie_tokenizer.py
@@ -1,6 +1,5 @@
-from typing import Any, List, Text, Optional, Dict
+from typing import Any, List, Text
 
-from rasa.nlu.components import Component
 from rasa.nlu.config import RasaNLUModelConfig
 from rasa.nlu.tokenizers.tokenizer import Token, Tokenizer
 from rasa.nlu.training_data import Message, TrainingData
