+++ rasa/nlu/tokenizers/tokenizer.py
@@ -1,3 +1,5 @@
+import logging
+
 from typing import Text, List, Optional, Dict, Any
 
 from rasa.nlu.components import Component
@@ -6,6 +8,8 @@ from rasa.nlu.constants import (
     MESSAGE_TEXT_ATTRIBUTE,
     CLS_TOKEN,
 )
+
+logger = logging.getLogger(__name__)
 
 
 class Token(object):
@@ -26,10 +30,13 @@ class Tokenizer(Component):
     def __init__(self, component_config: Optional[Dict[Text, Any]] = None) -> None:
         super(Tokenizer, self).__init__(component_config)
 
-        if "use_cls_token" in self.component_config:
+        try:
             self.use_cls_token = self.component_config["use_cls_token"]
-        else:
-            self.use_cls_token = True
+        except KeyError:
+            raise KeyError(
+                "No default value for 'use_cls_token' was set. Please, "
+                "add it to the default dict of the tokenizer."
+            )
 
     def add_cls_token(
         self, tokens: List[Token], attribute: Text = MESSAGE_TEXT_ATTRIBUTE
