+++ tests/nlu/tokenizers/test_whitespace_tokenizer.py
@@ -1,17 +1,14 @@
-# -*- coding: utf-8 -*-
-
-from unittest.mock import patch
-
 from rasa.nlu.constants import CLS_TOKEN
 from rasa.nlu.training_data import TrainingData, Message
 from tests.nlu import utilities
-from rasa.nlu import training_data
 
 
 def test_whitespace():
     from rasa.nlu.tokenizers.whitespace_tokenizer import WhitespaceTokenizer
 
-    tk = WhitespaceTokenizer()
+    component_config = {"use_cls_token": False}
+
+    tk = WhitespaceTokenizer(component_config)
 
     assert [t.text for t in tk.tokenize("Forecast for lunch")] == [
         "Forecast",
@@ -98,7 +95,11 @@ def test_whitespace_cls_token():
 def test_whitespace_custom_intent_symbol():
     from rasa.nlu.tokenizers.whitespace_tokenizer import WhitespaceTokenizer
 
-    component_config = {"intent_tokenization_flag": True, "intent_split_symbol": "+"}
+    component_config = {
+        "intent_tokenization_flag": True,
+        "intent_split_symbol": "+",
+        "use_cls_token": False,
+    }
 
     tk = WhitespaceTokenizer(component_config)
 
@@ -116,7 +117,7 @@ def test_whitespace_custom_intent_symbol():
 def test_whitespace_with_case():
     from rasa.nlu.tokenizers.whitespace_tokenizer import WhitespaceTokenizer
 
-    component_config = {"case_sensitive": False}
+    component_config = {"case_sensitive": False, "use_cls_token": False}
     tk = WhitespaceTokenizer(component_config)
     assert [t.text for t in tk.tokenize("Forecast for LUNCH")] == [
         "forecast",
@@ -124,7 +125,7 @@ def test_whitespace_with_case():
         "lunch",
     ]
 
-    component_config = {"case_sensitive": True}
+    component_config = {"case_sensitive": True, "use_cls_token": False}
     tk = WhitespaceTokenizer(component_config)
     assert [t.text for t in tk.tokenize("Forecast for LUNCH")] == [
         "Forecast",
@@ -132,7 +133,7 @@ def test_whitespace_with_case():
         "LUNCH",
     ]
 
-    component_config = {}
+    component_config = {"use_cls_token": False}
     tk = WhitespaceTokenizer(component_config)
     assert [t.text for t in tk.tokenize("Forecast for LUNCH")] == [
         "Forecast",
@@ -140,7 +141,7 @@ def test_whitespace_with_case():
         "LUNCH",
     ]
 
-    component_config = {"case_sensitive": False}
+    component_config = {"case_sensitive": False, "use_cls_token": False}
     tk = WhitespaceTokenizer(component_config)
     message = Message("Forecast for LUNCH")
     tk.process(message)
@@ -181,148 +182,3 @@ def test_whitespace_with_case():
     assert examples[1].data.get("tokens")[0].text == "i"
     assert examples[1].data.get("tokens")[1].text == "want"
     assert examples[1].data.get("tokens")[2].text == "tacos"
-
-
-def test_spacy(spacy_nlp):
-    from rasa.nlu.tokenizers.spacy_tokenizer import SpacyTokenizer
-
-    tk = SpacyTokenizer()
-
-    text = "Forecast for lunch"
-    assert [t.text for t in tk.tokenize(spacy_nlp(text))] == [
-        "Forecast",
-        "for",
-        "lunch",
-    ]
-    assert [t.offset for t in tk.tokenize(spacy_nlp(text))] == [0, 9, 13]
-
-    text = "hey ńöñàśçií how're you?"
-    assert [t.text for t in tk.tokenize(spacy_nlp(text))] == [
-        "hey",
-        "ńöñàśçií",
-        "how",
-        "'re",
-        "you",
-        "?",
-    ]
-    assert [t.offset for t in tk.tokenize(spacy_nlp(text))] == [0, 4, 13, 16, 20, 23]
-
-
-def test_spacy_add_cls_token(spacy_nlp):
-    from rasa.nlu.tokenizers.spacy_tokenizer import SpacyTokenizer
-
-    component_config = {"use_cls_token": True}
-
-    tk = SpacyTokenizer(component_config)
-
-    text = "Forecast for lunch"
-    assert [t.text for t in tk.tokenize(spacy_nlp(text))] == [
-        "Forecast",
-        "for",
-        "lunch",
-        CLS_TOKEN,
-    ]
-    assert [t.offset for t in tk.tokenize(spacy_nlp(text))] == [0, 9, 13, 19]
-
-
-def test_spacy_intent_tokenizer(spacy_nlp_component):
-    from rasa.nlu.tokenizers.spacy_tokenizer import SpacyTokenizer
-
-    td = training_data.load_data("data/examples/rasa/demo-rasa.json")
-    spacy_nlp_component.train(td, config=None)
-    spacy_tokenizer = SpacyTokenizer()
-    spacy_tokenizer.train(td, config=None)
-
-    intent_tokens_exist = [
-        True if example.get("intent_tokens") is not None else False
-        for example in td.intent_examples
-    ]
-
-    # no intent tokens should have been set
-    assert not any(intent_tokens_exist)
-
-
-def test_mitie():
-    from rasa.nlu.tokenizers.mitie_tokenizer import MitieTokenizer
-
-    tk = MitieTokenizer()
-
-    text = "Forecast for lunch"
-    assert [t.text for t in tk.tokenize(text)] == ["Forecast", "for", "lunch"]
-    assert [t.offset for t in tk.tokenize(text)] == [0, 9, 13]
-
-    text = "hey ńöñàśçií how're you?"
-    assert [t.text for t in tk.tokenize(text)] == [
-        "hey",
-        "ńöñàśçií",
-        "how",
-        "'re",
-        "you",
-        "?",
-    ]
-    assert [t.offset for t in tk.tokenize(text)] == [0, 4, 13, 16, 20, 23]
-
-
-def test_mitie_add_cls_token():
-    from rasa.nlu.tokenizers.mitie_tokenizer import MitieTokenizer
-
-    component_config = {"use_cls_token": True}
-
-    tk = MitieTokenizer(component_config)
-
-    text = "Forecast for lunch"
-    assert [t.text for t in tk.tokenize(text)] == [
-        "Forecast",
-        "for",
-        "lunch",
-        CLS_TOKEN,
-    ]
-    assert [t.offset for t in tk.tokenize(text)] == [0, 9, 13, 19]
-
-
-def test_jieba():
-    from rasa.nlu.tokenizers.jieba_tokenizer import JiebaTokenizer
-
-    tk = JiebaTokenizer()
-
-    assert [t.text for t in tk.tokenize("我想去吃兰州拉面")] == ["我", "想", "去", "吃", "兰州", "拉面"]
-
-    assert [t.offset for t in tk.tokenize("我想去吃兰州拉面")] == [0, 1, 2, 3, 4, 6]
-
-    assert [t.text for t in tk.tokenize("Micheal你好吗？")] == ["Micheal", "你好", "吗", "？"]
-
-    assert [t.offset for t in tk.tokenize("Micheal你好吗？")] == [0, 7, 9, 10]
-
-
-def test_jieba_load_dictionary(tmpdir_factory):
-    from rasa.nlu.tokenizers.jieba_tokenizer import JiebaTokenizer
-
-    dictionary_path = tmpdir_factory.mktemp("jieba_custom_dictionary").strpath
-
-    component_config = {"dictionary_path": dictionary_path}
-
-    with patch.object(
-        JiebaTokenizer, "load_custom_dictionary", return_value=None
-    ) as mock_method:
-        tk = JiebaTokenizer(component_config)
-        tk.tokenize("")
-
-    mock_method.assert_called_once_with(dictionary_path)
-
-
-def test_jieba_add_cls_token():
-    from rasa.nlu.tokenizers.jieba_tokenizer import JiebaTokenizer
-
-    component_config = {"use_cls_token": True}
-
-    tk = JiebaTokenizer(component_config)
-
-    assert [t.text for t in tk.tokenize("Micheal你好吗？")] == [
-        "Micheal",
-        "你好",
-        "吗",
-        "？",
-        CLS_TOKEN,
-    ]
-
-    assert [t.offset for t in tk.tokenize("Micheal你好吗？")] == [0, 7, 9, 10, 12]
